+++ octomap/src/ScanGraph.cpp
@@ -457,7 +457,7 @@ namespace octomap {
           currentNode = new ScanNode();
           currentNode->scan = new Pointcloud();
 
-          float x, y, z, roll, pitch, yaw;
+          double x, y, z, roll, pitch, yaw;
           std::string tmp;
           ss >> tmp >> x >> y >> z >> roll >> pitch >> yaw;
           pose6d pose(x, y, z, roll, pitch, yaw);
@@ -469,7 +469,7 @@ namespace octomap {
             OCTOMAP_ERROR_STR("Error parsing log file, no Scan to add point to!");
             break;
           }
-          float x, y, z;
+          double x, y, z;
           ss >> x >> y >> z;
 
           //std::cout << "Point "<< x << "," <<y <<"," <<z << " found.\n";
